fix(governance): rename React component utils.tsx to Utils.tsx (#PR4B)
