+++ apps/web/src/components/user/post-ad/steps/common/Utils.tsx
