+++ apps/web/src/components/user/post-ad/steps/listing-details/TitleSection.tsx
@@ -9,7 +9,7 @@ import { Button } from "@esparex/ui";
 import { Loader2 } from "@/icons/IconRegistry";
 import { MAX_AD_TITLE_CHARS } from "@esparex/contracts";
 import { AdPayload as PostAdFormData } from "@/schemas/adPayload.schema";
-import { getNestedFieldMeta, CharCounter } from "../common/utils";
+import { getNestedFieldMeta, CharCounter } from "../common/Utils";
 
 export function TitleSection() {
     const { register } = useFormContext<PostAdFormData>();
